Refactor: 장르 조회 로직 개선
